src/main/java/com/practice/javapractice2022/repository/MemberJpaRepository.java
package com.practice.javapractice2022.repository;

import com.practice.javapractice2022.entity.Member;
import org.springframework.stereotype.Repository;

import javax.persistence.EntityManager;
import javax.persistence.PersistenceContext;

@Repository
public class MemberJpaRepository {

    @PersistenceContext
    private EntityManager em; // jpa 사용을 위해 필요함

    public Member save(Member member) {
        em.persist(member);
        return member;
    }

    public Member find(Long id) {
        return em.find(Member.class, id);
    }
}
