src/test/java/com/practice/javapractice2022/repository/MemberJpaRepositoryTest.java
package com.practice.javapractice2022.repository;

import com.practice.javapractice2022.entity.Member;
import org.assertj.core.api.Assertions;
import org.junit.jupiter.api.Test;
import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.boot.test.context.SpringBootTest;
import org.springframework.test.annotation.Rollback;
import org.springframework.transaction.annotation.Transactional;

import static org.junit.jupiter.api.Assertions.*;
import static org.assertj.core.api.Assertions.*;

@SpringBootTest // junit5 에선 runwith뭐시기 필요없음
@Transactional // 트랜잭션에 필수요소이다. method 별로, class별로 설정 가능하다.
// @Rollback(false) // 테스트 데이터를 롤백하지 않도록 설정 할 수 있다.
class MemberJpaRepositoryTest {
    @Autowired
    MemberJpaRepository memberJpaRepository;

    @Test
    public void testMember() {
        Member member = new Member("member1");
        Member savedMember = memberJpaRepository.save(member);
        Member findMember = memberJpaRepository.find(savedMember.getId());
        assertThat(findMember.getId()).isEqualTo(member.getId());
        assertThat(findMember.getUsername()).isEqualTo(member.getUsername());
        assertThat(findMember).isEqualTo(member);
    }
}